requirements.txt
python-dotenv
openai==1.10.0
langchain==0.1.4
langchain-community==0.0.16
sqlalchemy==2.0.24
pypdf==4.0.1
chromadb==0.4.22
tiktoken==0.5.2
langchain-openai==0.0.5
panel==1.3.8
Pygments==2.15.1
anthropic==0.20.0